code/requirements.txt
pandas>=1.5.0
rank-bm25>=0.2.2
python-dotenv>=0.21.0
tqdm>=4.64.0
anthropic>=0.9.0
beautifulsoup4>=4.9.0
requests>=2.28.0
